Added some Exceptions

+++ src/App/Component/Film/Application/UseCase/Actor/DeleteActorUseCase.php
@@ -5,6 +5,7 @@ namespace App\Component\Film\Application\UseCase\Actor;
 use Doctrine\ORM\EntityManager;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use App\Bundle\FilmBundle\EventSubscriber\DeleteCache;
+use App\Component\Film\Application\Exception\InvalidActorIdException;
 
 class DeleteActorUseCase
 {
@@ -19,6 +20,10 @@ class DeleteActorUseCase
 
     public function execute(int $actorId)
     {
+        if(empty($actorId)) {
+            throw InvalidActorIdException::empty();
+        }
+
         $actor = $this->entityManager->getReference('\App\Component\Film\Domain\Actor', $actorId);
 
         $this->entityManager->remove($actor);
