+++ src/App/Component/Film/Application/UseCase/Film/DeleteFilmUseCase.php
@@ -5,6 +5,7 @@ namespace App\Component\Film\Application\UseCase\Film;
 use Doctrine\ORM\EntityManager;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use App\Bundle\FilmBundle\EventSubscriber\DeleteCache;
+use App\Component\Film\Application\Exception\InvalidFilmIdException;
 
 class DeleteFilmUseCase
 {
@@ -19,6 +20,10 @@ class DeleteFilmUseCase
 
     public function execute(int $filmId)
     {
+        if(empty($filmId)) {
+            throw InvalidFilmIdException::empty();
+        }
+
         $film = $this->entityManager->getReference('\App\Component\Film\Domain\Film', $filmId);
 
         $this->entityManager->remove($film);
